Parametrize a few of the states
- Parametrize hibernate, allow choice between hybrid-sleep,
suspend-then-hybernate and hibernate-only. The latter is the default.
- Parametrize mem-sleep-default. Allow switching between
s2idle and deep as it seems deep may not be the right
choice for everyone.
- Parametrize grub timeout to allow for disabling the timeout
completely, *which is not default since it's dangerous.*

Unrelated:

- Only enable service in fingerprint-reader since it isn't
a persistent service.
- Add polkit file for allowing hibernate. This should allow
people to get GNOME extension that gives them a Hibernate
item in the power menu that works.